Fix available GDC compilers.
The oldest available GDC binary is 4.8.2, based on 2.064.2. This is
very old (several versions before the latest C-based DMD frontend)
and there is no real point supporting such a version.
The latest available GDC binary is 6.3.0, but this is not available
on Travis. It looks that GDC 5.2.0 is the latest available binary
on Travis.