Another attempt to fix issue #115.
Root cause of the problem is that the used base data type for reading
an integer is always `ulong`. Converting a (huge) value to a signed,
possible shorter data type results in the onversion exception.

Solution is to use the correct type. This eliminates almost all calls
to the `std.conv.to` function.

All credits for the fix go to @timotheecour Timothee Cour as he found
the problem and created the first solution.